backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
openai>=1.3.8
twilio==8.10.0
requests==2.31.0
aiohttp==3.9.1
cors==1.0.1
# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
joblib>=1.3.0
apscheduler>=3.10.0
# ML & NLP
torch
transformers
sentence-transformers
scikit-learn
xgboost
# Vector database
chromadb
